Redraw main menus at 10fps
This is a workaround for an SDL bug on Wayland, in which resize events
are delayed until the next window update (GL buffer swap); by redrawing
the screen, we ensure all resizes take effect within a fraction of a
second. Window resizing continues to be smooth on X11 and on SDL versions
where the bug is patched.

This also ameliorates a problem where we do not correctly specify the window
size when we are creating a FULLSCREEN_DESKTOP window on startup/restart.