Migrate to SDL2 (#40)
While much of this change was straightforward, there are a few
exceptions:
* The conversion to the SDL_Renderer API introduced a surface/texture
  distinction; blits currently create temporary textures, costing an
  extra memory copy.
* Most of the video mode selection apparatus has been removed, leaving
  only a windowed/fullscreen option. In windowed mode, the window can
  be resized by the user; in fullscreen mode, the default
  (full-resolution) display mode is kept. As a result, the minimap and
  data panes in the bottom right corner may be too small on high
  resolution monitors.
* The main menu appears to be affected by an SDL bug on Wayland, where,
  in the absence of frequent repaints, resize events are delayed.
* The SDL2 now distinguishes between scancodes (positions on a keyboard)
  and keycodes (the resulting symbols, keyboard layout dependent.)
  Someone using a non-QWERTY keyboard should decide which commands should
  be linked to a keyboard position, and which should be linked to a
  mnemonic.

* Reorganize window resize handling

Now initVideo is only called to initialize video, and there
are separate methods to set a window size/fullscreen and to
update the OpenGL state when the event loop sees the resize.

* Use 'desktop' fullscreen instead of trying to change mode

Support for 'true' fullscreen with mode changes has limitations and
a few annoying side effects (like not always switching the video mode
back), so with this change Lincity-NG now uses SDL_FULLSCREEN_DESKTOP.

Since Lincity-NG often uses fixed sizes for icons and text, on some
displays when using fullscreen mode the text and icons will be very small.

* Redraw main menus at 10fps

This is a workaround for an SDL bug on Wayland, in which resize events
are delayed until the next window update (GL buffer swap); by redrawing
the screen, we ensure all resizes take effect within a fraction of a
second. Window resizing continues to be smooth on X11 and on SDL versions
where the bug is patched.

This also ameliorates a problem where we do not correctly specify the window
size when we are creating a FULLSCREEN_DESKTOP window on startup/restart.