remove help page references to an old powerscoal behavior
Remove de/el/pt_BR powerscoal help page references to an old
powerscoal behavior.

Actual text removed (in English) is:
Power lines will be connected to the northwest corner of the power plant,
transport routes on the two fields next to it. See example:
<img src="images/help/powerscoal-example.png" halign="center"/>
The example above shows a coal-fired power plant with all its connection
points used. On the left is a road, above it are two transmission lines.
On the right a railway line is connected to another transport connection.
To toggle the view of tall buildings press <b>h</b>.

Commit 1a19fc17ff079b5932c725c35ce4c13e490f71c6 did the same on the en version.

Also remove now unused powerscoal-example.png file.